+++ app/_layout.jsx
@@ -18,7 +18,7 @@ function RootLayoutNav() {
 
   useEffect(() => {
     if (!loading && !user) {
-      router.replace('/login');
+      router.replace('./index');
     }
   }, [user, loading]);
 
@@ -73,10 +73,10 @@ export default function RootLayout() {
         <Stack.Screen name="circuits" options={{ headerShown: false }} />
         <Stack.Screen name="chess" options={{ headerShown: false }} />
         <Stack.Screen name="csc" options={{ headerShown: false }} />
-        <Stack.Screen name="symbioses" options={{ headerShown: false }} />
+        <Stack.Screen name="symbiosis" options={{ headerShown: false }} />
         <Stack.Screen name="access" options={{ headerShown: false }} />
         <Stack.Screen name="storm" options={{ headerShown: false }} />
-        <Stack.Screen name="AdminDashboard" options={{ headerShown: false }} />
+        <Stack.Screen name="admin" options={{ headerShown: false }} />
         <Stack.Screen name="+not-found" />
       </Stack>
       <StatusBar style="auto" />
